# swagger/swagger_output.json
{
  "swagger": "2.0",
  "securityDefinitions": {
    "bearerAuth": {
      "name": "authorization",
      "in": "header",
      "type": "apiKey",
      "description": "JWT Authorization header"
    }
  },
  "security": [{ "bearerAuth": [] }],
  "info": {
    "title": "Cesta Solidária API",
    "description": "Api Cesta Solidária com node, express e mongodb",
    "version": "1.0.3",
    "contact": {
      "name": "Cesta Solidária",
      "email": "cartallumapp@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/api/v1/instituicoes/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "$ref": "#/definitions/Instituicao"
              }
            }
          },
          "404": {
            "description": "Nenhum documento encontrado para o id fornecido"
          }
        }
      },
      "post": {
        "description": "Cria instituição",
        "parameters": [
          {
            "name": "instituicao",
            "in": "body",
            "description": "Informações de instituição",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Instituicao"
            }
          }
        ],
        "responses": {
          "responses": {
            "200": {
              "description": "success",
              "schema": {
                "$ref": "#/definitions/Instituicao"
              }
            },
            "500": {
              "description": "ValidatorError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Local": {
      "type": "object",
      "properties": {
        "coordinates": {
          "type": "array"
        }
      }
    },
...
    "EnderecosInstituicao": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "logradouro": {
          "type": "string"
        },
        "bairro": {
          "type": "string"
        },
        "numero": {
          "type": "string"
        },
        "complemento": {
          "type": "string"
        },
        "pontoReferencia": {
          "type": "string"
        },
        "cep": {
          "type": "string"
        },
        "pais": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        },
        "local": {
          "type": "string"
        },
        "instituicao": {
          "type": "string"
        }
      }
    },
    "Instituicao": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "nome": {
          "type": "string",
          "example": "Instituição Matriz"
        },
        "sigla": {
          "type": "string",
          "example": "IM"
        },
        "tipo": {
          "type": "string",
          "description": "aceita os valores [matriz, instituicao]"
        },
        "telefone": {
          "type": "string"
        },
        "verificado": {
          "type": "boolean"
        },
        "verificador": {
          "type": "string"
        },
        "criadoEm": {
          "type": "string"
        },
        "local": {
          "type": "object",
          "$ref": "#/definitions/Local"
        },
        "endereco": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/EnderecosInstituicao"
          }
        }
      }
    },
...
  }
}